feat(ui): refactor model detail page and improve routing structure
