+++ ui/src/locales/en/shared.json
@@ -11,6 +11,12 @@
     "projectPlaceholder": "Select a project",
     "projectTooltip": "All models/datasets in a private project are private, and all models/datasets in a public project are public."
   },
+  "revisionSelect": {
+    "group": {
+      "branch": "Branch",
+      "tag": "Tag"
+    }
+  },
   "copyValueButton": {
     "copy": "Copy",
     "copied": "Copied"
